+++ src/main/java/net/servicestack/idea/NativeTypesLanguage.java
@@ -8,5 +8,6 @@ public enum NativeTypesLanguage {
     Kotlin,
     Dart,
     TypeScriptConcrete,
-    TypeScript
+    TypeScript,
+    Python
 }